Create reusable search container
- Search from homepage and navbar
- Custom sized input
